Decouple SupplyParameterFromQuery
