+++ src/Components/WebAssembly/WebAssembly/src/Hosting/WebAssemblyHostBuilder.cs
@@ -264,6 +264,6 @@ public sealed class WebAssemblyHostBuilder
         });
         Services.AddSingleton<FormDataProvider, DefaultFormDataProvider>();
         Services.AddSingleton<IFormValueSupplier, WebAssemblyFormValueSupplier>();
-        Services.AddSingleton<CascadingModelBindingProvider, CascadingQueryModelBindingProvider>();
+        Services.AddSupplyValueFromQueryProvider();
     }
 }
